Move arg parsing to utils
